requirements.txt
asgiref==3.4.1
djangify==1.0.0
Django==3.2.8
django-ckeditor==6.1.0
django-js-asset==1.2.2
Pillow==8.4.0
pytz==2021.3
sqlparse==0.4.2
